Update poissondistribution.tex
The indication of exercise 1.1.10 being an actual exercise was missing. Added 'show that' since this is the general language used in the book when a proof is required. Possibly the exercise would be even clearer if (1), (2), (3), (4) was added before each part of the exercise and correspondingly this numbering was added to the solutions. The solutions could be clearer if the proof of (1) was then added to the solutions, however, this proof is clearly trivial and hence this could be deemed unnecessary.