Update poissondistribution.tex
Concering that missing ). Not sure where I went wrong in that you couldn't find it. I've made the same pull request this time, but I'll add what I changed here:

In line 639 we originally have:
We see that the thinned stream is Poisson with parameter $\lambda p$. (For notational ease, we left out the $t$, otherwise it is $P(\lambda t p)$.

We should have:
We see that the thinned stream is Poisson with parameter $\lambda p$. (For notational ease, we left out the $t$, otherwise it is $P(\lambda t p)$.)

I hope you can find it now, sorry for the inconvenience.