Update kendall.tex
Just fixed a lot of inconsistencies. Also added a comma where I thought it was appropriate. I thought about making multiple pull requests for this to make it clear that it's multiple changes but to be honest that'd be incredibly obnoxious. Also, on the capitalization of words after colons:

"In British English, the first letter after a colon is capitalized only if it's a proper noun or an acronym; in American English, the first word after a colon is sometimes capitalized if it begins a complete sentence."
Source: https://www.grammarly.com/blog/capitalization-after-colons/

Personally I think using a capital letter looks better in this case.