Update open_single_class.tex
deleted the transpose sign, since I believe that the standing vector is just (x,y,z) and not (x,y,z)'. If it is not the right way to denote the standing vector this way, then the vectors given in line 122, namely w=(x,y,0) and v=(x,y,z) need a transpose sign.